metabase: Move _database_psycopg2 to root metabase folder
It makes no real sense to keep it in the management commands with an
underscore. It's a real usable module with an API.
